requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# zoo/main.py: 10
efficientnet == 1.1.0

# zoo/main.py: 4,8,86
keras == 2.4.3

# zoo/main.py: 2
matplotlib == 3.3.0

# zoo/main.py: 3
numpy == 1.13.3

# zoo/main.py: 11
omelette == 1.1.6

# zoo/main.py: 5
opencv-contrib-python == 4.3.0.36

# zoo/main.py: 5
opencv-contrib-python-headless == 4.3.0.36

# zoo/main.py: 5
opencv-openvino-contrib-python == 4.1.1.26

# zoo/main.py: 5
opencv-python == 4.3.0.36

# zoo/main.py: 5
opencv-python-aarch64 == 3.3.0.1

# zoo/main.py: 5
opencv-python-headless == 4.3.0.36

# zoo/main.py: 5
opencv-python-inference-engine == 4.4.0.0

# zoo/main.py: 5
opencvz-contrib-python == 4.1.1.26

# zoo/main.py: 1
pandas == 1.1.0

# zoo/main.py: 11
pillow == 7.2.0

# zoo/main.py: 11
pillow-simd == 7.0.0.post3

# zoo/main.py: 9
streamlit == 0.64.0

# zoo/main.py: 6
tensorflow == 2.3.0
